created startup, options structure
