+++ local.settings.json
@@ -5,10 +5,6 @@
         "FUNCTIONS_WORKER_RUNTIME": "dotnet",
         "AccountName": "devstoreaccount1",
         "Uri": "http://127.0.0.1:10002/devstoreaccount1",
-        "TableItem": "items",
-        "TableUser": "users",
-        "TableFavorites": "favorites",
-        "BlobImage": "images",
         "BlobUri": "http://127.0.0.1:10000/devstoreaccount1/images",
         "AccountKey": "Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==",
         "ConnectionString": "AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;DefaultEndpointsProtocol=http;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;"
